+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support for VMware vSphere 7.x to Oracle Cloud VMware Solution added
 - Support for Cost Analysis: Forecasting added
 - Support for OCI Vulnerability Scanning Service
+- Support for artifacts container image datasource
 
 ## 4.19.0 (March 24, 2021)